+++ Judge/Judge.Model/Contests/IContestsRepository.cs
@@ -5,6 +5,6 @@ namespace Judge.Model.Contests
     public interface IContestsRepository
     {
         IEnumerable<Contest> GetList();
-        Contest Get(long id);
+        Contest Get(int id);
     }
 }
